Add ability to merge all duplicate groups
I had way too many duplicates to manage manually.

This commit adds an action to the small arrow action bar of the
"Find Duplicates" plugin called "Merge All Groups" that, when
clicked, results in an iteration through all of the displayed
duplicate groups. On each iteration, Calibre's `Edit Metadata`
action's `merge_groups` method is called, which will display
an appropriate confirmation dialog box depending on the nature
of the selected group.

In order for this to work, the user must first trigger a view
displaying grouped duplicates, by clicking the "Find Duplicates"
button on the main Calibre toolbar.

Add docs